Rename manifestSelectors to rootStateSelectors
