+++ packages/interbit-covenant-tools/src/rootCovenant/index.js
@@ -1,10 +1,10 @@
 // © 2018 BTL GROUP LTD -  This package is licensed under the MIT license https://opensource.org/licenses/MIT
 const manifestCovenant = require('./manifestCovenant')
-const manifestSelectors = require('./manifestSelectors')
+const rootStateSelectors = require('./rootStateSelectors')
 const rootCovenant = require('./rootCovenant')
 
 module.exports = {
   manifestCovenant,
-  manifestSelectors,
+  rootStateSelectors,
   rootCovenant
 }
